Cleanup.
- Fixes type hint errors for PHP 7.4.
- Updates + completes docblocks.
- Cleans up logic and improves readability.
- Improves error output depending on number of errors to be displayed.

+++ src/console/controllers/ExampleTemplatesController.php
@@ -12,7 +12,6 @@ use craft\commerce\console\Controller;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Console;
 use craft\helpers\FileHelper;
-use GuzzleHttp\Exception\GuzzleException;
 use yii\base\ErrorException;
 use yii\base\Exception;
 use craft\helpers\Html;
@@ -32,10 +31,10 @@ class ExampleTemplatesController extends Controller
     public $defaultAction = 'generate';
 
     /**
-     * @var string Name of the folder the templates will copy into
+     * @var ?string Name of the target folder the templates will be copied to.
      * @since 3.3
      */
-    public string $folderName;
+    public string $folderName = '';
 
     /**
      * @var bool Whether to overwrite an existing folder. Must be passed if a folder with that name already exists.
@@ -44,10 +43,10 @@ class ExampleTemplatesController extends Controller
     public bool $overwrite = false;
 
     /**
-     * @var bool Whether to use HTMX
+     * @var ?bool Whether to use htmx
      * @since 3.3
      */
-    public bool $useHtmx;
+    public ?bool $useHtmx = null;
 
     /**
      * @var bool Whether to generate and copy to the example-templates build folder (used by Craft Commerce developers)
@@ -56,10 +55,10 @@ class ExampleTemplatesController extends Controller
     public bool $devBuild = false;
 
     /**
-     * @var string The base color for the generated example templates.
+     * @var ?string The base color for the generated example templates.
      * Possible values are: red, yellow, green, blue, indigo, purple, or pink.
      */
-    public string $baseColor;
+    public ?string $baseColor = null;
 
     /**
      * @var array
@@ -85,7 +84,7 @@ class ExampleTemplatesController extends Controller
     }
 
     /**
-     * Generate and copy the example templates.
+     * Generates and copies the example templates.
      *
      * @return int
      * @throws ErrorException
@@ -103,19 +102,19 @@ class ExampleTemplatesController extends Controller
         $slash = DIRECTORY_SEPARATOR;
         $pathService = Craft::$app->getPath();
         $templatesPath = $this->_getTemplatesPath();
+        $tempDestination = null;
 
         $exampleTemplatesSource = FileHelper::normalizePath(
             $pathService->getVendorPath() . '/craftcms/commerce/example-templates/src/shop'
         );
 
-        if ($this->folderName) {
+        if (!empty($this->folderName)) {
             $folderName = $this->folderName;
         } else {
             $this->stdout('A folder will be copied to your templates directory.' . PHP_EOL);
             $folderName = $this->prompt('Choose folder name:', ['required' => true, 'default' => 'shop']);
         }
 
-        // Use htmx
         if ($this->useHtmx === null) {
             $this->useHtmx = $this->confirm('Use htmx for forms and links?', true);
         }
@@ -137,13 +136,17 @@ class ExampleTemplatesController extends Controller
             // Create a temporary directory to hold the copy of the templates before we replace variables
             $tempDestination = $pathService->getTempPath() . $slash . 'commerce_example_templates_' . md5(uniqid(mt_rand(), true));
             // Copy the templates to the temporary directory
-            FileHelper::copyDirectory($exampleTemplatesSource, $tempDestination, ['recursive' => true, 'copyEmptyDirectories' => true]);
+            FileHelper::copyDirectory(
+                $exampleTemplatesSource,
+                $tempDestination,
+                ['recursive' => true, 'copyEmptyDirectories' => true]
+            );
 
-            // Find all text files in which we want to replace [[ ]] notation.
+            // Find all text files in which we want to replace [[ ]] notation
             $files = FileHelper::findFiles($tempDestination, [
                 'only' => ['*.twig', '*.html', '*.svg', '*.css'],
             ]);
-            // Set the [[ ]] notion variables and write the files
+            // Set the [[ ]] notation variables and write the files
             foreach ($files as $file) {
                 $fileContents = file_get_contents($file);
                 $fileContents = str_replace(
@@ -158,6 +161,10 @@ class ExampleTemplatesController extends Controller
             $errors[] = $e->getCode() . ' ' . $e->getMessage();
         }
 
+        if (!is_dir($tempDestination)) {
+            $errors[] = 'Could not generate templates.';
+        }
+
         if (!empty($errors)) {
             return $this->_returnErrors($errors);
         }
@@ -165,19 +172,16 @@ class ExampleTemplatesController extends Controller
         // New source is our temp directory ready for copying to site templates
         $source = $tempDestination;
 
-        // If this is a dev build, copy them to the build folder
         if ($this->devBuild) {
-            $destination = FileHelper::normalizePath(Craft::getAlias('@vendor') . '/craftcms/commerce/example-templates/dist/' . $this->folderName);
-        }
-
-        // If this is not a dev build, copy them to the templates folder
-        if (!$this->devBuild) {
-
+            // If this is a dev build, copy them to the build folder
+            $destination = FileHelper::normalizePath(
+                Craft::getAlias('@vendor') . '/craftcms/commerce/example-templates/dist/' . $this->folderName
+            );
+        } else {
+            // If this is not a dev build, copy them to the templates folder
             if (!$templatesPath) {
                 $errors[] = 'Can not determine the site template path.';
-            }
-
-            if ($templatesPath && !FileHelper::isWritable($templatesPath)) {
+            } else if (!FileHelper::isWritable($templatesPath)) {
                 $errors[] = 'Site template path is not writable.';
             }
 
@@ -188,25 +192,24 @@ class ExampleTemplatesController extends Controller
             $destination = $templatesPath . $slash . $folderName;
         }
 
-        $alreadyExists = is_dir($destination);
-        if ($alreadyExists && !$this->overwrite) {
-            $errors[] = 'Template folder "' . $folderName . '" already exists. Set the `overwrite` param to `true` if you want to replace it.';
+        $destinationExists = is_dir($destination);
+
+        if ($destinationExists && $this->overwrite) {
+            // We’re allowed to overwrite templates, and we’ve got valid source and destination folders
+            $this->stdout('Overwriting ...' . PHP_EOL, Console::FG_YELLOW);
+            FileHelper::removeDirectory($destination);
+        } else if ($destinationExists && !$this->overwrite) {
+            // A target folder’s been specified that already exists, but we’re not supposed to overwrite it
+            $errors[] = 'Template folder “' . $folderName . '” already exists. Replace it by setting the `overwrite` param to `true`.';
             return $this->_returnErrors($errors);
         }
 
-        if (is_dir($destination) && is_dir($source)) {
-            if ($this->overwrite) {
-                $this->stdout('Overwriting ...' . PHP_EOL, Console::FG_YELLOW);
-                FileHelper::removeDirectory($destination);
-            }
-        }
-        if (!is_dir($destination) && is_dir($source)) {
-            try {
-                $this->stdout('Copying ...' . PHP_EOL, Console::FG_YELLOW);
-                FileHelper::copyDirectory($source, $destination, ['recursive' => true, 'copyEmptyDirectories' => true]);
-            } catch (\Exception $e) {
-                $errors[] = $e->getMessage();
-            }
+        // Now let’s try and copy that template directory
+        try {
+            $this->stdout('Copying ...' . PHP_EOL, Console::FG_YELLOW);
+            FileHelper::copyDirectory($source, $destination, ['recursive' => true, 'copyEmptyDirectories' => true]);
+        } catch (\Exception $e) {
+            $errors[] = $e->getMessage();
         }
 
         if (!empty($errors)) {
@@ -218,6 +221,10 @@ class ExampleTemplatesController extends Controller
         return ExitCode::OK;
     }
 
+    /**
+     * Adds CSS key-value replacements to the array, where the key is our special `[[ ]]` template notation and
+     * the value is what it’ll be replaced with.
+     */
     private function _addCssClassesToReplacementData(): void
     {
         $mainColor = $this->baseColor ?: $this->select('Base Tailwind CSS color:', array_combine($this->_colors, $this->_colors));
@@ -238,6 +245,10 @@ class ExampleTemplatesController extends Controller
         ]);
     }
 
+    /**
+     * Adds external resource key-value replacements to the array, where the key is our special `[[ ]]` template
+     * notation and the value is what it’ll be replaced with.
+     */
     private function _addResourceAssetsToReplacementData(): void
     {
         $resourceTags = [
@@ -255,25 +266,38 @@ class ExampleTemplatesController extends Controller
     }
 
     /**
-     * @param array $errors
+     * Formats and outputs errors and exits.
+     *
+     * @param string[] $errors Error strings to be shown to the user
      * @return int
      */
     private function _returnErrors(array $errors): int
     {
-        $this->stderr('Error(s):' . PHP_EOL . '    - ' . implode(PHP_EOL . '    - ', $errors) . PHP_EOL, Console::FG_RED);
+        if (count($errors) === 1) {
+            $this->stderr('Error: ' . array_shift($errors) . PHP_EOL, Console::FG_RED);
+        } else {
+            $this->stderr(
+                'Errors:' . PHP_EOL . '    - ' . implode(PHP_EOL . '    - ', $errors) . PHP_EOL,
+                Console::FG_RED
+            );
+        }
+
         return ExitCode::USAGE;
     }
 
     /**
-     * @return string
+     * Returns the relevant site base template path.
+     *
+     * @return string The sites’s base template path
      * @throws Exception
      */
     private function _getTemplatesPath(): string
     {
-        $originalMode = Craft::$app->getView()->getTemplateMode();
-        Craft::$app->getView()->setTemplateMode(\craft\web\View::TEMPLATE_MODE_SITE);
-        $templatesPath = Craft::$app->getView()->getTemplatesPath();
-        Craft::$app->getView()->setTemplateMode($originalMode);
+        $view = Craft::$app->getView();
+        $originalMode = $view->getTemplateMode();
+        $view->setTemplateMode(\craft\web\View::TEMPLATE_MODE_SITE);
+        $templatesPath = $view->getTemplatesPath();
+        $view->setTemplateMode($originalMode);
         return $templatesPath;
     }
 }
